Content pass: two new low-level zones, XP curve fix, and four endgame zones (100-500)

+++ src/main/java/com/perso/T4C/helper/PlayerStateDto.java
@@ -24,8 +24,8 @@ public class PlayerStateDto {
   public int maxMana;
   public int mana;
   public int level;
-  public int currentXp;
-  public int xpToNextLevel;
+  public long currentXp;
+  public long xpToNextLevel;
   public int statPoints;
   public int skillPoints;
   public int rebirthCount;
